Add automatic geocoding based on street address, courtesy of Yahoo.
The way this works is that we install handlers on the address input
form elements, and attempt to geocode on either a "change" event
or when typing has paused for 3 seconds or so.  (The latter works
by setting a timeout on each keypress, cancelling the previous
timeout if one was already set.  It's not very pretty, but it works.)

Geocoding is currently done by lookup in the "geo.placefinder"
pseudo-table offered by Yahoo Query Language, querying Yahoo
straight from the browser.  This gets throttled by Yahoo at a
few thousand queries a day, but it's hard to forsee how a single
browser running this code is likely to reach that limit.