Properly note location of mobile "spots" index, even after redirects.
This works by setting the data-url property.  Note that this also
requires some tomfoolery around the (re)initialization of the map
if the page had already been loaded.  If we just feed the same <div>
to Leaflet that it had already used, it will complain that "Map
container is already initialized"; accordingly, we delete and
recreate (a common workaround, pace the mailing list).

To add to the confusion, jQmobile 1.2.0 will apparently, in some
circumstances, put two pages with the same ID in the DOM, so we
need to recreate the map in the right copy, which we detect by
looking at the target of the "pageinit" event.  This is quite
likely a bug, but a workaround is the easiest thing for now.