After saving a spot, display a map centered there, name popped up.
This involves a surprisingly awkward amount of machinery, mostly concerned
with feeding jQuery mobile unique IDs and data-urls for each centered map.
(Which might not be the ideal thing anyway, but altering a cached map in
the DOM is more voodoo than I want to contemplate right now.)  So, we
have .../spots/[id]/show-on-map, which renders HTML to render a map
centered on that spot.

(What actually happens is that the setup_map JS gets an extra argument,
which is the ID of the spot to highlight; it then picks that out of the
list, centers the map there, and arranges to display the popup with the
name.  The show_on_map Ruby in the controller actually does query for
the spot; this might be useful later as a hook for access control, or
to give us more of its attributes than just the ID.)