Consistent scheme for mobile_page_id and mobile_page_url.
Just does the obvious thing --- uses request.path to initialize both,
cleaning up punctuation for the ID.  Which should be fine so long as
POSTs always redirect, so there's no ambiguity about whether we're
displaying the result of show or edit (which, by Rails REST conventions,
would be show and update on the same URL).

The usual reason for POST redirection is so that a browser-level reload
after you're done won't re-POST.  This actually applies here, but it is
an extra HTTP request (meaning a couple of extra TCP-level round trips).

jQmobile actually provides ways to fake this out; send a data-url and
matching page ID for the redirect target.  Should consider, but this
sets up sensible defaults, and edits are not expected to be all that
common here...